+++ experiments/operational_domain/critical_temperature_domain_bestagon.cpp
@@ -24,7 +24,6 @@
 #include <fiction/technology/sidb/simulation/engine.hpp>
 #include <fiction/technology/sidb/simulation/logic/is_operational.hpp>
 #include <fiction/technology/sidb/simulation/logic/operational_domain.hpp>
-#include <fiction/types.hpp>
 
 #include <fmt/format.h>